Write the Markdown of Chimera build 50 alpha docs
This commit will introduce the Markdown equivalent of the Chimera build 50 alpha documentation, by Kavawuvi. It does so by updating the README.md to contain everything that has been written [here](https://docs.google.com/document/d/1OmkwbJTLmuVEqJGTtkLlbRpbdZ_QBHoj86MzC7mXM_g/edit#).

There are minor differences in terms of content - differences being typo fixes and wording changes.

In terms of presentation, the significant difference is that bullet points rather than tables are used for the commands' syntax and descriptions. This is due to the nightmare of dealing with complex tables in Markdown that contain multi-line & merged cells.

The following sections are present in this commit:

* Introductory, Requirements, Getting Started, Credits & Links
* All commands sections - Visuals, Xbox, Interpolation, Enhancements, Debug
* Building instructions from the initial version of the README.md